Satisfy governance for source observability
