+++ docs/product/prd/prd-42-source-governance-and-defaults.md
@@ -38,7 +38,7 @@ The product needs a narrow high-signal source mix. If catalog additions, fallbac
 - `src/lib/demo-data.ts` owns `MVP_DEFAULT_PUBLIC_SOURCE_IDS` and resolves the public default source set intentionally.
 - `src/adapters/donors/registry.ts` owns `DEFAULT_DONOR_FEED_IDS` and resolves no-argument donor fallback defaults intentionally.
 - `src/adapters/donors/registry.ts` also owns `PROBATIONARY_RUNTIME_FEED_IDS`, a separate governed runtime path for explicitly approved source trials. This is not an MVP default list and does not alter donor fallback defaults.
-- `src/lib/observability/runtime-source-resolution.ts` builds a safe runtime source-resolution snapshot with governed source IDs only.
+- `src/lib/observability/pipeline-run.ts` builds a safe runtime source-resolution snapshot with governed source IDs only.
 - `src/lib/pipeline/ingestion/index.ts` emits an ID-only structured log for runtime source resolution and attaches the same snapshot to pipeline-run metadata for local/test inspection.
 - `src/app/sources/page.tsx` presents catalog state as optional/importable rather than default ingestion.
 
@@ -66,7 +66,7 @@ The product needs a narrow high-signal source mix. If catalog additions, fallbac
   - `src/lib/source-policy.ts`
   - `src/lib/demo-data.ts`
   - `src/adapters/donors/registry.ts`
-  - `src/lib/observability/runtime-source-resolution.ts`
+  - `src/lib/observability/pipeline-run.ts`
   - `src/lib/pipeline/ingestion/index.ts`
   - `docs/engineering/change-records/source-catalog-cleanup-bbc-cnbc.md`
   - `docs/engineering/change-records/source-onboarding-model.md`
